Collected auth related code under auth namespace
With auth related code in one namespace, we get a nice separation of
concerns and the auth implementation details are abstracted away from
other modules.

Not sure if placing the auth in the access layer, i.e. UsersWeb
,namespace is correct decision, but it is easy to move it elsewhere if
necessary.
